+++ src/features/TextToImage/client/exports.ts
@@ -1,3 +1,3 @@
 export {
   default as SDXL,
-} from '@/features/TextToImage/models/SDXL/client/index.ts';
+} from '@/features/TextToImage/models/SDXL/client/exports';
